night updates

+++ Server/MaliMaisonApi/Migrations/CameraDbContextModelSnapshot.cs
@@ -35,8 +35,8 @@ namespace MaliMaisonApi.Migrations
                     b.Property<string>("Name")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<long>("Price")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Price")
+                        .HasColumnType("int");
 
                     b.Property<int>("Stock")
                         .HasColumnType("int");
